+++ UniBiblio/Models/Prenotazionieffettuatesaleview.cs
@@ -6,7 +6,7 @@ namespace UniBiblio.Models
     public partial class Prenotazionieffettuatesaleview
     {
         public ulong IdPrenotazione { get; set; }
-        public int? IdUtente { get; set; }
+        public ulong? IdUtente { get; set; }
         public string EmailUtente { get; set; } = null!;
         public string NomeSala { get; set; } = null!;
         public DateOnly DataPrenotazione { get; set; }
